Convert `ContinueReceiving` to an async method.
After a few hours of debugging, it looks like the ContinueReceiving method
that was changed to adapt to Azure Service Bus three was blocking the
thread and causing a lot of intermittent connection issues.

This patch is not necessarily a fix for the problem, it was just the
fastest way to get around the symptoms.